requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
sqlalchemy==2.0.23
Pillow==10.1.0 